extract interprocess communication logic from example and added it to frameworks/shared
