+++ examples/custom/RandomForest_standalone/__init__.py
@@ -1,8 +1,13 @@
+import os
+
+from amlb.resources import config as rconfig
 from amlb.utils import call_script_in_same_dir
 
+__path__.append(os.path.join(rconfig().root_dir, 'frameworks', 'shared'))
 
-def setup(*args):
-    call_script_in_same_dir(__file__, "setup.sh", *args)
+
+def setup(*args, **kwargs):
+    call_script_in_same_dir(__file__, "setup.sh", rconfig().root_dir, *args, **kwargs)
 
 
 def run(*args, **kwargs):
